Big refactor of run command into seperate class.
