+++ src/Aspire.Cli/Commands/BaseCommand.cs
@@ -1,0 +1,16 @@
+// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+
+using System.CommandLine;
+
+namespace Aspire.Cli.Commmands;
+
+internal abstract class BaseCommand : Command
+{
+    protected BaseCommand(string name, string description) : base(name, description)
+    {
+        SetAction(ExecuteAsync);
+    }
+
+    protected abstract Task<int> ExecuteAsync(ParseResult parseResult, CancellationToken cancellationToken);
+}
